refactor(MP-13): Move AddSupplier.tsx to proper folder.
